+++ src/theme/components/structuredContent.ts
@@ -1,25 +1,33 @@
 const StructuredContent = {
-  parts: ["image"],
+  parts: ["image", "imageContainer"],
   baseStyle: ({ imagePosition }) => {
-    const styles = {};
+    const styles: { maxWidth?: string } = {};
     if (imagePosition === "center") {
-      (styles as any).maxWidth = "100%";
+      styles.maxWidth = "100%";
     }
     return {
-      image: {
+      imageContainer: {
         float: imagePosition !== "center" ? imagePosition : undefined,
+        marginBottom: "m",
         marginLeft:
           imagePosition === "center"
             ? null
             : imagePosition === "left"
             ? "0"
-            : "s",
+            : "m",
         marginRight:
           imagePosition === "center"
             ? null
             : imagePosition === "left"
-            ? "s"
+            ? "m"
             : "0",
+        width: imagePosition !== "center" ? "100%" : undefined,
+        ...styles,
+      },
+      image: {
+        // To handle the case when the image does not have a container,
+        // e.g. when there is no image caption or credit.
+        float: imagePosition !== "center" ? imagePosition : undefined,
         ...styles,
       },
     };